Started on merchant NPC, new map for first level
I started work on the merchant, he currently is used for development purposes and carries an infinite amount of all items for free, his inventory will be tweaked and he will be used to sell basic items to the player.

I bought an asset pack for a 16x16 tile map and started mapping out the first level (building zone), the forest.